refactor(MapState tests): extract fixture
