+++ src/apps/Map/tests/fixtures/planningsResponse.json
@@ -1,0 +1,50 @@
+{
+  "count": 245,
+  "next": "https://fixmyberlin.de/api/projects?page=2&page_size=1",
+  "previous": null,
+  "results": [
+    {
+      "id": 4,
+      "url": "https://fixmyberlin.de/api/projects/4",
+      "project_key": "FK-020",
+      "title": "Neue Fahrradstraße",
+      "description": "Die Glogauer Straße ist als Zubringer zum Görlitzer Park und Durchfahrt zur Oberbaumbrücke eine stark befahrene Radstrecke. Die Straße soll daher als Fahrradstraße umgwidmet werden. Die Machbarkeit soll im Jahr 2019 geprüft und bei einem positivem Ergebnis voraussichtlich in 2019 umgesetzt werden.",
+      "short_description": "Die Glogauer Straße wird zu einer Fahrradstraße.",
+      "category": "bike street",
+      "street_name": "Glogauer Straße",
+      "borough": "Friedrichshain-Kreuzberg",
+      "side": 2,
+      "costs": null,
+      "draft_submitted": null,
+      "construction_started": null,
+      "construction_completed": "unbekannt",
+      "construction_completed_date": null,
+      "phase": "draft",
+      "responsible": "Bezirksamt Friedrichshain-Kreuzberg",
+      "external_url": null,
+      "cross_section": null,
+      "faq": [],
+      "geometry": {
+        "type": "LineString",
+        "coordinates": [
+          [13.435653786621, 52.4916709861841],
+          [13.4357418335253, 52.4917627197358],
+          [13.4374143916796, 52.4934570093144],
+          [13.4389228394647, 52.4949934633041]
+        ]
+      },
+      "center": {
+        "type": "Point",
+        "coordinates": [13.4374143916796, 52.4934570093144]
+      },
+      "length": 431.0,
+      "photos": [
+        {
+          "copyright": "Photo by Anthony Ginsbrook",
+          "src": "https://fmb-aws-bucket.s3.amazonaws.com/photos/Platzhalter_anthony-ginsbrook-225252-unsplash.jpg"
+        }
+      ],
+      "likes": 4
+    }
+  ]
+}
